+++ myems-api/core/store.py
@@ -436,6 +436,19 @@ class StoreItem:
                                    title='API.BAD_REQUEST',
                                    description='API.THERE_IS_RELATION_WITH_VIRTUAL_METERS')
 
+        # check relation with command
+        cursor.execute(" SELECT command_id "
+                       " FROM tbl_stores_commands "
+                       " WHERE store_id = %s ",
+                       (id_,))
+        rows_commands = cursor.fetchall()
+        if rows_commands is not None and len(rows_commands) > 0:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_400,
+                                   title='API.BAD_REQUEST',
+                                   description='API.THERE_IS_RELATION_WITH_COMMANDS')
+
         cursor.execute(" DELETE FROM tbl_stores WHERE id = %s ", (id_,))
         cnx.commit()
 
@@ -1691,6 +1704,177 @@ class StoreWorkingCalendarItem:
                                    description='API.STORE_WORKING_CALENDAR_RELATION_NOT_FOUND')
 
         cursor.execute(" DELETE FROM tbl_stores_working_calendars WHERE store_id = %s AND working_calendar_id = %s ", (id_, wcid))
+        cnx.commit()
+
+        cursor.close()
+        cnx.close()
+
+        resp.status = falcon.HTTP_204
+
+
+class StoreCommandCollection:
+    @staticmethod
+    def __init__():
+        """Initializes Class"""
+        pass
+
+    @staticmethod
+    def on_options(req, resp, id_):
+        resp.status = falcon.HTTP_200
+
+    @staticmethod
+    def on_get(req, resp, id_):
+        if not id_.isdigit() or int(id_) <= 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_STORE_ID')
+
+        cnx = mysql.connector.connect(**config.myems_system_db)
+        cursor = cnx.cursor()
+
+        cursor.execute(" SELECT name "
+                       " FROM tbl_stores "
+                       " WHERE id = %s ", (id_,))
+        if cursor.fetchone() is None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                   description='API.STORE_NOT_FOUND')
+
+        query = (" SELECT c.id, c.name, c.uuid "
+                 " FROM tbl_stores s, tbl_stores_commands sc, tbl_commands c "
+                 " WHERE sc.store_id = s.id AND c.id = sc.command_id AND s.id = %s "
+                 " ORDER BY c.id ")
+        cursor.execute(query, (id_,))
+        rows = cursor.fetchall()
+
+        result = list()
+        if rows is not None and len(rows) > 0:
+            for row in rows:
+                meta_result = {"id": row[0], "name": row[1], "uuid": row[2]}
+                result.append(meta_result)
+
+        resp.text = json.dumps(result)
+
+    @staticmethod
+    @user_logger
+    def on_post(req, resp, id_):
+        """Handles POST requests"""
+        access_control(req)
+        try:
+            raw_json = req.stream.read().decode('utf-8')
+        except Exception as ex:
+            raise falcon.HTTPError(status=falcon.HTTP_400,
+                                   title='API.BAD_REQUEST',
+                                   description='API.FAILED_TO_READ_REQUEST_STREAM')
+
+        if not id_.isdigit() or int(id_) <= 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_STORE_ID')
+
+        new_values = json.loads(raw_json)
+
+        if 'command_id' not in new_values['data'].keys() or \
+                not isinstance(new_values['data']['command_id'], int) or \
+                new_values['data']['command_id'] <= 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_COMMAND_ID')
+        command_id = new_values['data']['command_id']
+
+        cnx = mysql.connector.connect(**config.myems_system_db)
+        cursor = cnx.cursor()
+
+        cursor.execute(" SELECT name "
+                       " from tbl_stores "
+                       " WHERE id = %s ", (id_,))
+        if cursor.fetchone() is None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                   description='API.STORE_NOT_FOUND')
+
+        cursor.execute(" SELECT name "
+                       " FROM tbl_commands "
+                       " WHERE id = %s ", (command_id,))
+        if cursor.fetchone() is None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                   description='API.COMMAND_NOT_FOUND')
+
+        query = (" SELECT id " 
+                 " FROM tbl_stores_commands "
+                 " WHERE store_id = %s AND command_id = %s")
+        cursor.execute(query, (id_, command_id,))
+        if cursor.fetchone() is not None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.ERROR',
+                                   description='API.STORE_COMMAND_RELATION_EXISTS')
+
+        add_row = (" INSERT INTO tbl_stores_commands (store_id, command_id) "
+                   " VALUES (%s, %s) ")
+        cursor.execute(add_row, (id_, command_id,))
+        cnx.commit()
+        cursor.close()
+        cnx.close()
+
+        resp.status = falcon.HTTP_201
+        resp.location = '/stores/' + str(id_) + '/commands/' + str(command_id)
+
+
+class StoreCommandItem:
+    @staticmethod
+    def __init__():
+        """Initializes Class"""
+        pass
+
+    @staticmethod
+    def on_options(req, resp, id_, cid):
+        resp.status = falcon.HTTP_200
+
+    @staticmethod
+    @user_logger
+    def on_delete(req, resp, id_, cid):
+        access_control(req)
+        if not id_.isdigit() or int(id_) <= 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_STORE_ID')
+
+        if not cid.isdigit() or int(cid) <= 0:
+            raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
+                                   description='API.INVALID_COMMAND_ID')
+
+        cnx = mysql.connector.connect(**config.myems_system_db)
+        cursor = cnx.cursor()
+
+        cursor.execute(" SELECT name "
+                       " FROM tbl_stores "
+                       " WHERE id = %s ", (id_,))
+        if cursor.fetchone() is None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                   description='API.STORE_NOT_FOUND')
+
+        cursor.execute(" SELECT name "
+                       " FROM tbl_commands "
+                       " WHERE id = %s ", (cid,))
+        if cursor.fetchone() is None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                   description='API.COMMAND_NOT_FOUND')
+
+        cursor.execute(" SELECT id "
+                       " FROM tbl_stores_commands "
+                       " WHERE store_id = %s AND command_id = %s ", (id_, cid))
+        if cursor.fetchone() is None:
+            cursor.close()
+            cnx.close()
+            raise falcon.HTTPError(status=falcon.HTTP_404, title='API.NOT_FOUND',
+                                   description='API.STORE_COMMAND_RELATION_NOT_FOUND')
+
+        cursor.execute(" DELETE FROM tbl_stores_commands WHERE store_id = %s AND command_id = %s ", (id_, cid))
         cnx.commit()
 
         cursor.close()